Rename colorScheme to color-scheme for consistency
